+++ test/remote-svf-to-gltf.js
@@ -19,9 +19,9 @@ async function run (urn, outputDir) {
         const modelDerivativeClient = new ModelDerivativeClient(auth);
         const helper = new ManifestHelper(await modelDerivativeClient.getManifest(urn));
         const derivatives = helper.search({ type: 'resource', role: 'graphics' });
-        const writer0 = new GltfWriter(path.join(outputDir, 'gltf-nodedup'), { deduplicate: false, binary: false, compress: false, log: console.log });
-        const writer1 = new GltfWriter(path.join(outputDir, 'gltf'), { deduplicate: true, binary: false, compress: false, log: console.log });
-        const writer2 = new GltfWriter(path.join(outputDir, 'glb-draco'), { deduplicate: true, binary: true, compress: true, log: console.log });
+        const writer0 = new GltfWriter(path.join(outputDir, 'gltf-nodedup'), { deduplicate: false, skipUnusedUvs: false, binary: false, compress: false, log: console.log });
+        const writer1 = new GltfWriter(path.join(outputDir, 'gltf'), { deduplicate: true, skipUnusedUvs: true, binary: false, compress: false, log: console.log });
+        const writer2 = new GltfWriter(path.join(outputDir, 'glb-draco'), { deduplicate: true, skipUnusedUvs: true, binary: true, compress: true, log: console.log });
         for (const derivative of derivatives.filter(d => d.mime === 'application/autodesk-svf')) {
             const reader = await SvfReader.FromDerivativeService(urn, derivative.guid, auth);
             const svf = await reader.read({ log: console.log });