+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added
+  - Skipping texture UVs when there's no material using them
+
 ## [0.6.3] - 2019-10-17
 
 - Changed